bpf: move L2_RESPONDER_MAP4 to l2_responder.h, cilium_l2_responder_v4
