+++ bpf/lib/maps.h
@@ -68,24 +68,6 @@ struct bpf_elf_map __section_maps cilium_calls = {
 };
 #endif /* SKIP_CALLS_MAP */
 
-struct l2_responder_v4_key {
-	__u32 ip4;
-	__u32 ifindex;
-};
-
-struct l2_responder_v4_stats {
-	__u64 responses_sent;
-};
-
-struct {
-	__uint(type, BPF_MAP_TYPE_HASH);
-	__type(key, struct l2_responder_v4_key);
-	__type(value, struct l2_responder_v4_stats);
-	__uint(pinning, LIBBPF_PIN_BY_NAME);
-	__uint(max_entries, L2_RESPONSER_MAP4_SIZE);
-	__uint(map_flags, BPF_F_NO_PREALLOC);
-} L2_RESPONDER_MAP4 __section_maps_btf;
-
 #ifdef ENABLE_SRV6
 # define SRV6_VRF_MAP(IP_FAMILY)				\
 struct {						\
