+++ bpf/node_config.h
@@ -142,7 +142,6 @@ volatile const __u64 __config_ROUTER_IP_2;
 #define SIGNAL_MAP test_cilium_signals
 #define POLICY_STATS_MAP_SIZE 200
 #define ENCRYPT_MAP test_cilium_encrypt_state
-#define L2_RESPONDER_MAP4 test_cilium_l2_responder_v4
 #define RATELIMIT_MAP test_cilium_ratelimit
 #define RATELIMIT_METRICS_MAP test_cilium_ratelimit_metrics
 #define TUNNEL_MAP test_cilium_tunnel_map
